pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plottrbot"
version = "0.1.0"
description = "Linux desktop controller for Plottrbot robot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Pillow>=10.4.0",
  "PySide6>=6.7.0",
  "pyserial>=3.5",
  "platformdirs>=4.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-qt>=4.4.0",
]

[project.scripts]
plottrbot = "plottrbot.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
